+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+* **CHANGED:** Deprecate the use of an array of chars for string literals; use `str` instead.
+
 ## [v0.2.0] - 2021-07-26
 
 * **CHANGED:** Merge the `goto` instruction with the `br` instruction.